pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "domo"
version = "0.1.0"
description = "Domo personal assistant and job workflow tools"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
  "pypdf",
  "pydantic",
  "PyYAML",
  "reportlab",
  "requests",
  "streamlit",
]

[project.scripts]
domo-job = "tools.job.main:main"

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
include = ["assistant*", "integrations*", "tools*", "workflows*"]
